Split main helper
Some tests (i.e. unit) do not require a full
integration, just access to some basics like
`loadFixtures()`, the configuration, db, etc.
